+++ ERPWebAPI.EL/Concrete/OHS/OHS_TestQuestionType.cs
@@ -8,7 +8,7 @@ namespace ERPWebAPI.EL.Concrete.OHS
     public class OHS_TestQuestionType : IEntity
     {
         [Key]
-        public int QuestionTypeId { get; set; }
+        public short QuestionTypeId { get; set; }
         public string QuestionType { get; set; }
         public string? UserEmployee { get; set; }
         public DateTime? TransactionDate { get; set; }
