+++ ERPWebAPI.EL/Concrete/SYS/SYS_PaperworkDepartment.cs
@@ -8,7 +8,7 @@ namespace ERPWebAPI.EL.Concrete.SYS
     public class SYS_PaperworkDepartment : IEntity
     {
         [Key]
-        public int PaperworkDepartmentId { get; set; }
+        public short PaperworkDepartmentId { get; set; }
         public string PaperworkDepartment { get; set; }
         public string PaperworkDepartmentCode { get; set; }
         public string? UserEmployee { get; set; }
